Update password_strenght_checker.py
This Python script assesses the strength of a password by analyzing its length, the presence of uppercase and lowercase letters, numbers, and special characters. It offers constructive feedback to guide users in crafting more robust passwords. The script is straightforward, interactive, and user-friendly, making it easy to enhance password security.